pyproject.toml
[tool.poetry]
name = "vespucci.ai"
version = "0.1.0"
description = "A Discord bot that integrates with Groq LLM and MCP servers"
authors = ["Your Name <your.email@example.com>"]
# readme = "README.md"  # Commented out until README.md is created
packages = [
    {include = "api"},
    {include = "client"},
    {include = "config"},
    {include = "discord"},
    {include = "models"},
    {include = "utils"},
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
discord-py = "^2.3.2"
groq = "^0.4.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"

httpx = "^0.27.0"
asyncio = "^3.4.3"
python-json-logger = "^2.0.7"
mcp = "^1.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.poetry.scripts]
start = "main:app"